+++ kjr_event/tests/__init__.py
@@ -1,3 +1,5 @@
 # -*- coding: utf-8 -*-
 from . import test_registration
+from . import test_event_document
+from . import test_event_meeting_info
 from . import test_event_seats_min